+++ apps/3d_rec_framework/include/pcl/apps/3d_rec_framework/pipeline/global_nn_recognizer_crh.h
@@ -53,7 +53,7 @@ namespace pcl
       typedef typename pcl::PointCloud<PointInT>::Ptr PointInTPtr;
       typedef typename pcl::PointCloud<PointInT>::ConstPtr ConstPointInTPtr;
 
-      typedef Distance<float> DistT;
+      typedef Distance<double> DistT;
       typedef Model<PointInT> ModelT;
       typedef pcl::PointCloud<pcl::Histogram<90> > CRHPointCloud;
 
@@ -78,7 +78,7 @@ namespace pcl
       int ICP_iterations_;
 
       bool noisify_;
-      float noise_;
+      double noise_;
 
       class flann_model
       {
@@ -86,18 +86,18 @@ namespace pcl
         ModelT model;
         int view_id;
         int descriptor_id;
-        std::vector<float> descr;
+        std::vector<double> descr;
       };
 
-      flann::Matrix<float> flann_data_;
+      flann::Matrix<double> flann_data_;
       flann::Index<DistT> * flann_index_;
       std::vector<flann_model> flann_models_;
 
 
       bool use_cache_;
-      std::map<std::pair<std::string, int>, Eigen::Matrix4f, std::less<std::pair<std::string, int> >, Eigen::aligned_allocator<std::pair<std::pair<
-          std::string, int>, Eigen::Matrix4f> > > poses_cache_;
-      std::map<std::pair<std::string, int>, Eigen::Vector3f > centroids_cache_;
+      std::map<std::pair<std::string, int>, Eigen::Matrix4d, std::less<std::pair<std::string, int> >, Eigen::aligned_allocator<std::pair<std::pair<
+          std::string, int>, Eigen::Matrix4d> > > poses_cache_;
+      std::map<std::pair<std::string, int>, Eigen::Vector3d > centroids_cache_;
 
       std::vector<int> indices_;
 
@@ -106,12 +106,12 @@ namespace pcl
       loadFeaturesAndCreateFLANN ();
 
       inline void
-      convertToFLANN (const std::vector<flann_model> &models, flann::Matrix<float> &data)
+      convertToFLANN (const std::vector<flann_model> &models, flann::Matrix<double> &data)
       {
         data.rows = models.size ();
         data.cols = models[0].descr.size (); // number of histogram bins
 
-        flann::Matrix<float> flann_data (new float[models.size () * models[0].descr.size ()], models.size (), models[0].descr.size ());
+        flann::Matrix<double> flann_data (new double[models.size () * models[0].descr.size ()], models.size (), models[0].descr.size ());
 
         for (size_t i = 0; i < data.rows; ++i)
           for (size_t j = 0; j < data.cols; ++j)
@@ -123,16 +123,16 @@ namespace pcl
       }
 
       void
-      nearestKSearch (flann::Index<DistT> * index, const flann_model &model, int k, flann::Matrix<int> &indices, flann::Matrix<float> &distances);
+      nearestKSearch (flann::Index<DistT> * index, const flann_model &model, int k, flann::Matrix<int> &indices, flann::Matrix<double> &distances);
 
       void
-      getPose (ModelT & model, int view_id, Eigen::Matrix4f & pose_matrix);
+      getPose (ModelT & model, int view_id, Eigen::Matrix4d & pose_matrix);
 
       void
       getCRH (ModelT & model, int view_id, int d_id, CRHPointCloud::Ptr & hist);
 
       void
-      getCentroid (ModelT & model, int view_id, int d_id, Eigen::Vector3f & centroid);
+      getCentroid (ModelT & model, int view_id, int d_id, Eigen::Vector3d & centroid);
 
       void
       getView (ModelT & model, int view_id, PointInTPtr & view);
@@ -140,7 +140,7 @@ namespace pcl
       int NN_;
 
       boost::shared_ptr<std::vector<ModelT> > models_;
-      boost::shared_ptr<std::vector<Eigen::Matrix4f, Eigen::aligned_allocator<Eigen::Matrix4f> > > transforms_;
+      boost::shared_ptr<std::vector<Eigen::Matrix4d, Eigen::aligned_allocator<Eigen::Matrix4d> > > transforms_;
 
     public:
 
@@ -155,7 +155,7 @@ namespace pcl
       {
       }
 
-      void setNoise(float n) {
+      void setNoise(double n) {
         noisify_ = true;
         noise_ = n;
       }
@@ -251,7 +251,7 @@ namespace pcl
        return models_;
       }
 
-      boost::shared_ptr<std::vector<Eigen::Matrix4f, Eigen::aligned_allocator<Eigen::Matrix4f> > >
+      boost::shared_ptr<std::vector<Eigen::Matrix4d, Eigen::aligned_allocator<Eigen::Matrix4d> > >
       getTransforms ()
       {
        return transforms_;
