+++ apps/3d_rec_framework/include/pcl/apps/3d_rec_framework/utils/persistence_utils.h
@@ -11,7 +11,7 @@ namespace pcl
     {
 
       inline bool
-      writeCentroidToFile (std::string file, Eigen::Vector3f & centroid)
+      writeCentroidToFile (std::string file, Eigen::Vector3d & centroid)
       {
         std::ofstream out (file.c_str ());
         if (!out)
@@ -27,7 +27,7 @@ namespace pcl
       }
 
       inline bool
-      getCentroidFromFile (std::string file, Eigen::Vector3f & centroid)
+      getCentroidFromFile (std::string file, Eigen::Vector3d & centroid)
       {
         std::ifstream in;
         in.open (file.c_str (), std::ifstream::in);
@@ -37,15 +37,15 @@ namespace pcl
         std::string line (linebuf);
         std::vector < std::string > strs;
         boost::split (strs, line, boost::is_any_of (" "));
-        centroid[0] = static_cast<float> (atof (strs[0].c_str ()));
-        centroid[1] = static_cast<float> (atof (strs[1].c_str ()));
-        centroid[2] = static_cast<float> (atof (strs[2].c_str ()));
+        centroid[0] = static_cast<double> (atof (strs[0].c_str ()));
+        centroid[1] = static_cast<double> (atof (strs[1].c_str ()));
+        centroid[2] = static_cast<double> (atof (strs[2].c_str ()));
 
         return true;
       }
 
       inline bool
-      writeMatrixToFile (std::string file, Eigen::Matrix4f & matrix)
+      writeMatrixToFile (std::string file, Eigen::Matrix4d & matrix)
       {
         std::ofstream out (file.c_str ());
         if (!out)
@@ -69,7 +69,7 @@ namespace pcl
       }
 
       inline bool
-      writeFloatToFile (std::string file, float value)
+      writeFloatToFile (std::string file, double value)
       {
         std::ofstream out (file.c_str ());
         if (!out)
@@ -100,7 +100,7 @@ namespace pcl
       }
 
       inline bool
-      readFloatFromFile (std::string dir, std::string file, float& value)
+      readFloatFromFile (std::string dir, std::string file, double& value)
       {
 
         std::vector < std::string > strs;
@@ -120,13 +120,13 @@ namespace pcl
 
         char linebuf[1024];
         in.getline (linebuf, 1024);
-        value = static_cast<float> (atof (linebuf));
+        value = static_cast<double> (atof (linebuf));
 
         return true;
       }
 
       inline bool
-      readFloatFromFile (std::string file, float& value)
+      readFloatFromFile (std::string file, double& value)
       {
 
         std::ifstream in;
@@ -134,13 +134,13 @@ namespace pcl
 
         char linebuf[1024];
         in.getline (linebuf, 1024);
-        value = static_cast<float> (atof (linebuf));
+        value = static_cast<double> (atof (linebuf));
 
         return true;
       }
 
       inline bool
-      readMatrixFromFile (std::string dir, std::string file, Eigen::Matrix4f & matrix)
+      readMatrixFromFile (std::string dir, std::string file, Eigen::Matrix4d & matrix)
       {
 
         //get the descriptor name from dir
@@ -184,14 +184,14 @@ namespace pcl
 
         for (int i = 0; i < 16; i++)
         {
-          matrix (i % 4, i / 4) = static_cast<float> (atof (strs_2[i].c_str ()));
+          matrix (i % 4, i / 4) = static_cast<double> (atof (strs_2[i].c_str ()));
         }
 
         return true;
       }
 
       inline bool
-      readMatrixFromFile (std::string file, Eigen::Matrix4f & matrix)
+      readMatrixFromFile (std::string file, Eigen::Matrix4d & matrix)
       {
 
         std::ifstream in;
@@ -205,14 +205,14 @@ namespace pcl
 
         for (int i = 0; i < 16; i++)
         {
-          matrix (i % 4, i / 4) = static_cast<float> (atof (strs_2[i].c_str ()));
+          matrix (i % 4, i / 4) = static_cast<double> (atof (strs_2[i].c_str ()));
         }
 
         return true;
       }
 
       inline bool
-      readMatrixFromFile2 (std::string file, Eigen::Matrix4f & matrix)
+      readMatrixFromFile2 (std::string file, Eigen::Matrix4d & matrix)
       {
 
         std::ifstream in;
@@ -226,7 +226,7 @@ namespace pcl
 
         for (int i = 0; i < 16; i++)
         {
-          matrix (i / 4, i % 4) = static_cast<float> (atof (strs_2[i].c_str ()));
+          matrix (i / 4, i % 4) = static_cast<double> (atof (strs_2[i].c_str ()));
         }
 
         return true;
